requirements.txt
blinker==1.4
certifi==2018.8.24
chardet==3.0.4
click==6.7
dnspython==1.15.0
elasticsearch==6.3.1
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
Flask-WhooshAlchemy==0.56
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pipenv==2018.7.1
py3dns==3.2.0
requests==2.19.1
SQLAlchemy==1.2.11
urllib3==1.23
validate-email==1.3
virtualenv==16.0.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
Whoosh==2.7.4
